Move source selector to OverlayControls
This move accomplishes a few things
1) It looks nicer
2) Can take advantage of the query error handling UI
3) Gives plenty of room for the dropdown
