+++ ui/src/style/chronograf.scss
@@ -50,6 +50,7 @@
 @import 'components/resizer';
 @import 'components/search-widget';
 @import 'components/source-indicator';
+@import 'components/source-selector';
 @import 'components/tables';
 
 // Pages
