+++ ui/src/style/components/source-selector.scss
@@ -1,0 +1,11 @@
+/*
+    Source Selector component styles
+    ----------------------------------------------------------------
+*/
+
+.source-selector {
+  display: flex;
+  align-items: center;
+  flex-wrap: nowrap;
+  flex: 1 0 0;
+}
